Update for code format and optimization
A big update is coming!
RealMute is a project started by me about one and a half years ago. At
that time, I was a programming novice who didn't know so many manners
of writing codes or so much knowledge about PHP. After my learning of
programming for more than one year, I have been aware of more rules
regarding code format and conventions, so I revisited this project and
made a lot of changes that reflected most of the knowledge I studied.